requirements.txt
aioredis==1.1.0
asgiref==2.3.2
async-timeout==2.0.1
attrs==18.1.0
autobahn==18.5.2
Automat==0.6.0
certifi==2018.4.16
channels==2.1.1
channels-redis==2.2.1
chardet==3.0.4
constantly==15.1.0
daphne==2.1.2
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.6
django-cors-headers==2.2.0
django-custom-user==0.7
djangorestframework==3.8.2
gunicorn==19.8.1
hiredis==0.2.0
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
msgpack==0.5.6
oauthlib==2.1.0
pipenv==2018.5.18
psycopg2==2.7.4
PyJWT==1.6.4
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==1.0.0
schedule==0.5.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
Twisted==18.4.0
txaio==2.10.0
urllib3==1.22
virtualenv==16.0.0
virtualenv-clone==0.3.0
whitenoise==3.3.1
zope.interface==4.5.0
